permissions are not checked in the views Fixes #236

+++ classified_ads/views.py
@@ -1,7 +1,6 @@
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.core.exceptions import PermissionDenied
 from django.http import JsonResponse
 from django.shortcuts import redirect, get_object_or_404
 from django.urls import reverse
@@ -10,7 +9,7 @@ from django.views import generic
 from django.views.decorators.csrf import csrf_exempt
 
 from classified_ads.forms import AdPhotoForm, ClassifiedAdForm
-from cm_main.utils import assert_request_is_ajax
+from cm_main.utils import assert_request_is_ajax, check_edit_permission
 
 from .models import AdPhoto, ClassifiedAd, Categories
 
@@ -45,8 +44,7 @@ class UpdateAdView(LoginRequiredMixin, generic.UpdateView):
         return reverse('classified_ads:detail', args=[self.object.pk])
 
     def get_context_data(self, **kwargs):
-        if self.request.user != self.get_object().owner:
-            raise PermissionDenied
+        check_edit_permission(self.request, self.get_object().owner)
         context = super().get_context_data(**kwargs)
         context['categories'] = Categories()
         context['photo_form'] = AdPhotoForm()
@@ -63,8 +61,7 @@ class DeleteAdView(LoginRequiredMixin, generic.View):
 
   def post(self, request, pk):
     ad = get_object_or_404(self.model, pk=pk)
-    if self.request.user != ad.owner:
-      raise PermissionDenied
+    check_edit_permission(request, ad.owner)
     ad.delete()
     messages.success(request, _("Ad \"%(title)s\" deleted") % {"title": ad.title})
     return redirect("classified_ads:list")
@@ -88,6 +85,7 @@ class AdPhotoAddView(LoginRequiredMixin, generic.View):
         form = AdPhotoForm(request.POST, self.request.FILES)
         if form.is_valid():
             form.instance.ad = get_object_or_404(ClassifiedAd, pk=self.kwargs['pk'])
+            check_edit_permission(request, form.instance.ad.owner)
             form.save()
             messages.success(self.request, _("Photo added successfully"))
         return redirect('classified_ads:update', pk=self.kwargs['pk'])
@@ -98,6 +96,7 @@ class AdPhotoAddView(LoginRequiredMixin, generic.View):
 def delete_photo(request, pk):
     assert_request_is_ajax(request)
     photo = get_object_or_404(AdPhoto, pk=pk)
+    check_edit_permission(request, photo.ad.owner)
     photo.delete()
     return JsonResponse({'success': True})
